+++ src/components/cards/crowdfunding-card/components/informations.js
@@ -33,6 +33,10 @@ var _colorsConfig = _interopRequireDefault(require("../../../../constants/colors
 
 var _screenConfig = require("../../../../constants/screen-config");
 
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
+
 var COMPONENT_GUTTER = (0, _typography.pxToRem)(10);
 
 var StyledContainer = _styledComponents.default.div.withConfig({
@@ -59,14 +63,14 @@ var StyledText = (0, _styledComponents.default)(_text.Text).withConfig({
   componentId: "qg8b01-4"
 })(["font-size:", ";"], (0, _typography.pxToRem)(12));
 
-var Informations =
-/*#__PURE__*/
-function (_PureComponent) {
+var Informations = /*#__PURE__*/function (_PureComponent) {
   (0, _inherits2.default)(Informations, _PureComponent);
+
+  var _super = _createSuper(Informations);
 
   function Informations() {
     (0, _classCallCheck2.default)(this, Informations);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(Informations).apply(this, arguments));
+    return _super.apply(this, arguments);
   }
 
   (0, _createClass2.default)(Informations, [{