+++ src/components/cards/simple-card/index.js
@@ -39,19 +39,23 @@ var _paragraph = require("./components/paragraph");
 
 var _colorsConfig = _interopRequireDefault(require("../../../constants/colors-config"));
 
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
+
 var ContainerStyle = _styledComponents.default.a.withConfig({
   displayName: "simple-card__ContainerStyle",
   componentId: "sc-1g5bppt-0"
 })(["line-height:1;position:relative;"]);
 
-var SimpleCardComponent =
-/*#__PURE__*/
-function (_Component) {
+var SimpleCardComponent = /*#__PURE__*/function (_Component) {
   (0, _inherits2.default)(SimpleCardComponent, _Component);
+
+  var _super = _createSuper(SimpleCardComponent);
 
   function SimpleCardComponent() {
     (0, _classCallCheck2.default)(this, SimpleCardComponent);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(SimpleCardComponent).apply(this, arguments));
+    return _super.apply(this, arguments);
   }
 
   (0, _createClass2.default)(SimpleCardComponent, [{