+++ src/components/cards/team-card/image.js
@@ -29,14 +29,18 @@ var _marger = require("../../layout/marger");
 
 var _colorsConfig = _interopRequireDefault(require("../../../constants/colors-config"));
 
-var TeamCardImage =
-/*#__PURE__*/
-function (_Component) {
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
+
+var TeamCardImage = /*#__PURE__*/function (_Component) {
   (0, _inherits2.default)(TeamCardImage, _Component);
+
+  var _super = _createSuper(TeamCardImage);
 
   function TeamCardImage() {
     (0, _classCallCheck2.default)(this, TeamCardImage);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(TeamCardImage).apply(this, arguments));
+    return _super.apply(this, arguments);
   }
 
   (0, _createClass2.default)(TeamCardImage, [{